Demographic Information of All Countries
In this project we used SQLite database to analyze CIA factbook data. This data contains demographic and geographic details of all the countries in the world. Objective of this project is to find out which countries top the list in the queries (what is the maximum/minimum population, which countries are densely populated? etc.) we are running.